hatena-agent-v2/requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
lxml==5.1.0
python-dotenv==1.0.0
chromadb==0.4.22
langchain==0.1.6
langchain-community==0.0.19
Pillow==10.2.0
aiohttp==3.9.3
pytest==8.0.0
pytest-asyncio==0.23.3
bingart